refactor: ruborist_adapter
